Creacion de trips con datos minimos (plan de viaje)

+++ database/migrations/tenant/2019_07_29_023902_drop_gps_field_of_trailer_box_table.php
@@ -13,9 +13,11 @@ class DropGpsFieldOfTrailerBoxTable extends Migration
      */
     public function up()
     {
-        Schema::table('trailer_boxes', function (Blueprint $table){
-            $table->dropColumn('gps');
-        });
+        if (Schema::hasColumn('trailer_boxes','gps')){
+            Schema::table('trailer_boxes', function (Blueprint $table){
+                $table->dropColumn('gps');
+            });
+        }
     }
 
     /**
